requirements.txt
# Maintainer: Leandro Michelino | ACE | leandro.michelino@oracle.com
oci>=2.173.0
streamlit>=1.36.0
python-dotenv>=1.0.1
pydantic>=2.7.0
pydantic-settings>=2.2.1
tenacity>=8.3.0
markdown>=3.6
pandas>=2.2.0
pypdf>=5.0.0
rich>=13.7.1
